Make comment display on submit of form
